+++ elide-core/src/main/java/com/yahoo/elide/ElideResponse.java
@@ -7,12 +7,15 @@ package com.yahoo.elide;
 
 import lombok.Getter;
 
+import java.util.Optional;
+
 /**
  * Elide response object.
  */
 public class ElideResponse {
     @Getter private final int responseCode;
     @Getter private final String body;
+    @Getter private final Optional<Throwable> failureReason;
 
     /**
      * Constructor.
@@ -20,8 +23,21 @@ public class ElideResponse {
      * @param responseCode HTTP response code
      * @param body returned body string
      */
+    @Deprecated
     public ElideResponse(int responseCode, String body) {
+        this(responseCode, body, null);
+    }
+
+    /**
+     * Constructor.
+     *
+     * @param responseCode HTTP response code
+     * @param body returned body string
+     * @param failureReason the reason the request failed
+     */
+    public ElideResponse(int responseCode, String body, Throwable failureReason) {
         this.responseCode = responseCode;
         this.body = body;
+        this.failureReason = Optional.ofNullable(failureReason);
     }
 }